points command: ApplicationCommand-only
